MigrationsCeleryStrikesAgain (Provide more useful info when older initializers are used and Migrations is enabled)
This change alters the way database initializers behave when it is detected that Migrations is enabled, especially when the database does not exist. The problem is that if Migrations is enabled but then one of the older initializers is used we decided that it should not create the database since, in general, Migrations should be used for this.

There were two problems with what we had before:
- If the user does Enable-Migrations and then hits F5 we want a useful message. As the code was before we just failed in a relatively obscure way because the database doesn't exist.
- There was a bug that meant the Drop... initializers would drop the database and then fail in an obscure way because the database does not exist.

The fix is to throw with a useful message if we know that the database doesn't exist but Migrations is enabled. Specifically, if a Migrations configuration is discovered, then:
 - If the database doesn't exist, then throw informing user to use Migrations to create the database
 - If the database exists but the model is out-of-date, then throw informing the user that the model is stale and Migrations should be used to update it. (This is the current behavior for CreateDatabaseIfNotExists anyway, and now makes sense for DropCreateDatabaseIfModelChanges as well.)
 - If the database exists and is up-to-date, then do nothing.
 - If the database exists but contains no MigrationHistory table, then do nothing (which is current behavior) except for DropCreateDatabaseIfModelChanges where it is expected that there will be metadata so we throw as before.

Note that "database exists" means that the database both exists and contains some table that can be mapped to an entity type. In other words, empty databases are considered databases that don't exist.

CreateDatabase also has the behavior that it will throw if Migrations is configured, which is the behavior introduced some time ago. This makes sense because DbMigrator should be used to create the database in this scenario.

Note that the change tries to make sure that an actual Exists check on the database does not happen more often than it did previously.

+++ src/EntityFramework/DropCreateDatabaseAlways`.cs
@@ -3,6 +3,7 @@
 namespace System.Data.Entity
 {
     using System.Data.Entity.Config;
+    using System.Data.Entity.Internal;
     using System.Data.Entity.Utilities;
 
     /// <summary>
@@ -14,6 +15,18 @@ namespace System.Data.Entity
     public class DropCreateDatabaseAlways<TContext> : IDatabaseInitializer<TContext>
         where TContext : DbContext
     {
+        private readonly MigrationsChecker _migrationsChecker;
+
+        public DropCreateDatabaseAlways()
+            : this(null)
+        {
+        }
+
+        internal DropCreateDatabaseAlways(MigrationsChecker migrationsChecker)
+        {
+            _migrationsChecker = migrationsChecker ?? new MigrationsChecker();
+        }
+
         #region Strategy implementation
 
         static DropCreateDatabaseAlways()
@@ -34,6 +47,13 @@ namespace System.Data.Entity
         public void InitializeDatabase(TContext context)
         {
             Check.NotNull(context, "context");
+
+            if (_migrationsChecker.IsMigrationsConfigured(
+                context.GetType(), 
+                () => new DatabaseTableChecker().AnyModelTableExists(context.InternalContext)))
+            {
+                return;
+            }
 
             context.Database.Delete();
             context.Database.Create();
